// src/app/models/messages.ts
// tslint:disable-next-line:class-name
export interface messages {
    iD: string;
    code: number;
    fromNumber: number;
    toNumber: number;
    messageID: string;
    chatID: string;
    dateSent: Date;
    readMessage: Date;
    timeMessage: Date;
    mediaEncode: string;
    userSent: number;
    nameSentMessage: string;
    status: string;
    typeMessage: string;
    replyId: string;
    toAnyOne: string;
    contentMessage: string;
}